Refactor GlucoDataService companion functions
