+++ common/src/main/java/de/michelinside/glucodatahandler/common/GlucoDataService.kt
@@ -1,43 +1,26 @@
 package de.michelinside.glucodatahandler.common
 
 import android.annotation.SuppressLint
-import android.app.AlarmManager
 import android.app.ForegroundServiceStartNotAllowedException
 import android.app.Notification
-import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
 import android.content.IntentFilter
 import android.content.SharedPreferences
 import android.content.pm.ServiceInfo
-import android.graphics.Color
 import android.os.Build
 import android.os.Bundle
-import android.provider.Settings
 import androidx.annotation.RequiresApi
 import com.google.android.gms.wearable.WearableListenerService
 import de.michelinside.glucodatahandler.common.database.dbAccess
 import de.michelinside.glucodatahandler.common.notification.ChannelType
 import de.michelinside.glucodatahandler.common.notification.Channels
-import de.michelinside.glucodatahandler.common.notifier.DataSource
 import de.michelinside.glucodatahandler.common.notifier.InternalNotifier
 import de.michelinside.glucodatahandler.common.notifier.NotifySource
-import de.michelinside.glucodatahandler.common.receiver.AAPSReceiver
-import de.michelinside.glucodatahandler.common.receiver.AidexBroadcastReceiver
 import de.michelinside.glucodatahandler.common.receiver.BatteryReceiver
 import de.michelinside.glucodatahandler.common.receiver.BroadcastServiceAPI
-import de.michelinside.glucodatahandler.common.receiver.DexcomBroadcastReceiver
-import de.michelinside.glucodatahandler.common.receiver.DiaboxReceiver
-import de.michelinside.glucodatahandler.common.receiver.GlucoseDataReceiver
-import de.michelinside.glucodatahandler.common.receiver.LibrePatchedReceiver
-import de.michelinside.glucodatahandler.common.receiver.NamedBroadcastReceiver
-import de.michelinside.glucodatahandler.common.receiver.NamedReceiver
-import de.michelinside.glucodatahandler.common.receiver.NotificationReceiver
-import de.michelinside.glucodatahandler.common.receiver.NsEmulatorReceiver
 import de.michelinside.glucodatahandler.common.receiver.ScreenEventReceiver
-import de.michelinside.glucodatahandler.common.receiver.XDripBroadcastReceiver
 import de.michelinside.glucodatahandler.common.tasks.BackgroundWorker
-import de.michelinside.glucodatahandler.common.tasks.LibreLinkSourceTask
 import de.michelinside.glucodatahandler.common.tasks.SourceTaskService
 import de.michelinside.glucodatahandler.common.tasks.TimeTaskService
 import de.michelinside.glucodatahandler.common.utils.GlucoDataUtils
@@ -46,9 +29,10 @@ import de.michelinside.glucodatahandler.common.utils.Log
 import de.michelinside.glucodatahandler.common.utils.PackageUtils
 import de.michelinside.glucodatahandler.common.utils.TextToSpeechUtils
 import de.michelinside.glucodatahandler.common.utils.Utils
-import java.util.Locale
+import de.michelinside.glucodatahandler.common.service.ReceiverManager
+import de.michelinside.glucodatahandler.common.service.StartupTrigger
+import de.michelinside.glucodatahandler.common.service.WearPhoneManager
 import androidx.core.content.edit
-import de.michelinside.glucodatahandler.common.notification.AlarmType
 
 
 enum class AppSource {
@@ -67,9 +51,6 @@ abstract class GlucoDataService(source: AppSource) : WearableListenerService(), 
     companion object {
         private val LOG_ID = "GDH.GlucoDataService"
         private var isForegroundService = false
-        @JvmStatic
-        @SuppressLint("StaticFieldLeak")
-        protected var connection: WearPhoneConnection? = null
         var startServiceReceiver: Class<*>? = null
         val foreground get() = isForegroundService
         const val NOTIFICATION_ID = 1234
@@ -141,7 +122,7 @@ abstract class GlucoDataService(source: AppSource) : WearableListenerService(), 
                     isRunning = true
                     if (!foreground && startServiceReceiver != null) {
                         // trigger also foreground alarm
-                        triggerStartService(context, startServiceReceiver!!)
+                        StartupTrigger.triggerStartService(context, startServiceReceiver!!)
                     }
                 } catch (exc: Exception) {
                     if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.S && exc is ForegroundServiceStartNotAllowedException) {
@@ -155,70 +136,13 @@ abstract class GlucoDataService(source: AppSource) : WearableListenerService(), 
                             context.startService(serviceIntent)
                         }
                         if(startServiceReceiver != null) {
-                            triggerStartService(context, startServiceReceiver!!)
+                            StartupTrigger.triggerStartService(context, startServiceReceiver!!)
                         }
                     } else {
                         Log.e(LOG_ID,"start exception: " + exc.message.toString() + "\n" + exc.stackTraceToString())
                         isRunning = false
                     }
                 }
-            }
-        }
-
-        private var alarmManager: AlarmManager? = null
-        private var alarmPendingIntent: PendingIntent? = null
-
-        private fun stopTrigger() {
-            try {
-                if(alarmManager != null && alarmPendingIntent != null) {
-                    Log.i(LOG_ID, "Stop trigger")
-                    alarmManager!!.cancel(alarmPendingIntent!!)
-                    alarmManager = null
-                    alarmPendingIntent = null
-                }
-            } catch (exc: Exception) {
-                Log.e(LOG_ID, "stopTrigger exception: " + exc.message.toString())
-            }
-        }
-
-        private fun triggerStartService(context: Context, receiver: Class<*>) {
-            try {
-                Log.i(LOG_ID, "Trigger start service - foreground: $foreground - alarm active: ${alarmManager != null && alarmPendingIntent != null}")
-                if(foreground || (alarmManager != null && alarmPendingIntent != null))
-                    return
-                alarmManager = context.getSystemService(ALARM_SERVICE) as AlarmManager
-                var hasExactAlarmPermission = true
-                if (!Utils.canScheduleExactAlarms(context)) {
-                    Log.d(LOG_ID, "Need permission to set exact alarm!")
-                    hasExactAlarmPermission = false
-                }
-                val intent = Intent(context, receiver)
-                intent.action = Constants.ACTION_START_FOREGROUND
-                intent.addFlags(Intent.FLAG_INCLUDE_STOPPED_PACKAGES)
-                alarmPendingIntent = PendingIntent.getBroadcast(
-                    context,
-                    911,
-                    intent,
-                    PendingIntent.FLAG_IMMUTABLE or PendingIntent.FLAG_CANCEL_CURRENT
-                )
-                val alarmTime = System.currentTimeMillis() + 1000
-                Log.i(LOG_ID, "Trigger alarm at ${Utils.getUiTimeStamp(alarmTime)} - exactAlarm: $hasExactAlarmPermission")
-                if (hasExactAlarmPermission) {
-                    alarmManager!!.setExactAndAllowWhileIdle(
-                        AlarmManager.RTC_WAKEUP,
-                        alarmTime,
-                        alarmPendingIntent!!
-                    )
-                } else {
-                    alarmManager!!.setAndAllowWhileIdle(
-                        AlarmManager.RTC_WAKEUP,
-                        alarmTime,
-                        alarmPendingIntent!!
-                    )
-                }
-            } catch (exc: Exception) {
-                Log.e(LOG_ID, "triggerStartService exception: " + exc.message.toString())
-                stopTrigger()
             }
         }
 
@@ -231,45 +155,6 @@ abstract class GlucoDataService(source: AppSource) : WearableListenerService(), 
             }
         }
 
-        fun checkForConnectedNodes(refreshDataOnly: Boolean = false) {
-            try {
-                Log.d(LOG_ID, "checkForConnectedNodes called for dataOnly=$refreshDataOnly - connection: ${connection!=null}")
-                if (connection!=null) {
-                    if(!refreshDataOnly)
-                        connection!!.checkForConnectedNodes(true)
-                    else
-                        connection!!.checkForNodesWithoutData()
-                }
-            } catch (exc: Exception) {
-                Log.e(
-                    LOG_ID,
-                    "checkForConnectedNodes exception: " + exc.message.toString()
-                )
-            }
-        }
-
-        fun resetWearPhoneConnection() {
-            try {
-                if (connection != null) {
-                    connection!!.resetConnection()
-                }
-            } catch (exc: Exception) {
-                Log.e(LOG_ID,"resetWearPhoneConnection exception: " + exc.message.toString())
-            }
-        }
-
-        fun sendCommand(command: Command, extras: Bundle? = null) {
-            try {
-                if (connection!=null)
-                    connection!!.sendCommand(command, extras)
-            } catch (exc: Exception) {
-                Log.e(
-                    LOG_ID,
-                    "sendCommand exception: " + exc.message.toString()
-                )
-            }
-        }
-
         fun resetDB() {
             try {
                 Log.w(LOG_ID, "reset database called!")
@@ -278,303 +163,18 @@ abstract class GlucoDataService(source: AppSource) : WearableListenerService(), 
                 GlucoseStatistics.reset()
                 ReceiveData.reset(context!!)
                 if (appSource == AppSource.PHONE_APP)
-                    sendCommand(Command.CLEAN_UP_DB)
+                    WearPhoneManager.sendCommand(Command.CLEAN_UP_DB)
             } catch (exc: Exception) {
                 Log.e(LOG_ID, "resetDB exception: " + exc.message.toString())
             }
         }
 
-        fun getWearPhoneConnection(): WearPhoneConnection? {
-            return connection
-        }
-
-        private var glucoDataReceiver: GlucoseDataReceiver? = null
-        private var xDripReceiver: XDripBroadcastReceiver?  = null
-        private var librePatchedReceiver: LibrePatchedReceiver?  = null
-        private var aapsReceiver: AAPSReceiver?  = null
-        private var dexcomReceiver: DexcomBroadcastReceiver? = null
-        private var nsEmulatorReceiver: NsEmulatorReceiver? = null
-        private var diaboxReceiver: DiaboxReceiver? = null
-        private var notificationReceiver: NotificationReceiver? = null
-        private var aidexReceiver: AidexBroadcastReceiver? = null
-        private val registeredReceivers = mutableSetOf<String>()
-
-        @SuppressLint("UnspecifiedRegisterReceiverFlag")
-        fun registerReceiver(context: Context, receiver: NamedReceiver, filter: IntentFilter): Boolean {
-            Log.i(LOG_ID, "Register receiver ${receiver.getName()} for $receiver on $context")
-            try {
-                if (receiver is NamedBroadcastReceiver) {
-                    PackageUtils.registerReceiver(context, receiver, filter)
-                }
-                registeredReceivers.add(receiver.getName())
-                return true
-            } catch (exc: Exception) {
-                Log.e(LOG_ID, "registerReceiver exception: " + exc.toString())
-            }
-            return false
-        }
-
-        fun unregisterReceiver(context: Context, receiver: NamedReceiver?) {
-            try {
-                if (receiver != null) {
-                    Log.i(LOG_ID, "Unregister receiver ${receiver.getName()} on $context")
-                    registeredReceivers.remove(receiver.getName())
-                    if (receiver is NamedBroadcastReceiver) {
-                        context.unregisterReceiver(receiver)
-                    }
-                }
-            } catch (exc: Exception) {
-                Log.e(LOG_ID, "unregisterReceiver exception: " + exc.toString())
-            }
-        }
-
-        fun isRegistered(receiver: NamedReceiver): Boolean {
-            return registeredReceivers.contains(receiver.getName())
-        }
-
-        fun updateSourceReceiver(context: Context, key: String? = null) {
-            Log.d(LOG_ID, "Register receiver")
-            try {
-                val sharedPref = context.getSharedPreferences(Constants.SHARED_PREF_TAG, MODE_PRIVATE)
-                if(key.isNullOrEmpty() || key == Constants.SHARED_PREF_SOURCE_JUGGLUCO_ENABLED) {
-                    if (sharedPref.getBoolean(Constants.SHARED_PREF_SOURCE_JUGGLUCO_ENABLED, true)) {
-                        if(glucoDataReceiver == null) {
-                            glucoDataReceiver = GlucoseDataReceiver()
-                            if(!registerReceiver(context, glucoDataReceiver!!, IntentFilter("glucodata.Minute")))
-                                glucoDataReceiver = null
-                        }
-                    } else if (glucoDataReceiver != null) {
-                        unregisterReceiver(context, glucoDataReceiver)
-                        glucoDataReceiver = null
-                    }
-                }
-
-                if(key.isNullOrEmpty() || key == Constants.SHARED_PREF_SOURCE_XDRIP_ENABLED) {
-                    if (sharedPref.getBoolean(Constants.SHARED_PREF_SOURCE_XDRIP_ENABLED, true)) {
-                        if(xDripReceiver == null) {
-                            xDripReceiver = XDripBroadcastReceiver()
-                            if(!registerReceiver(context, xDripReceiver!!, IntentFilter("com.eveningoutpost.dexdrip.BgEstimate")))
-                                xDripReceiver = null
-                        }
-                    } else if (xDripReceiver != null) {
-                        unregisterReceiver(context, xDripReceiver)
-                        xDripReceiver = null
-                    }
-                }
-
-                if(key.isNullOrEmpty() || key == Constants.SHARED_PREF_SOURCE_AAPS_ENABLED) {
-                    if (sharedPref.getBoolean(Constants.SHARED_PREF_SOURCE_AAPS_ENABLED, true)) {
-                        if(aapsReceiver == null) {
-                            aapsReceiver = AAPSReceiver()
-                            if(!registerReceiver(context, aapsReceiver!!, IntentFilter(Intents.AAPS_BROADCAST_ACTION)))
-                                aapsReceiver = null
-                        }
-                    } else if (aapsReceiver != null) {
-                        unregisterReceiver(context, aapsReceiver)
-                        aapsReceiver = null
-                    }
-                }
-
-                if(key.isNullOrEmpty() || key == Constants.SHARED_PREF_SOURCE_BYODA_ENABLED) {
-                    if (sharedPref.getBoolean(Constants.SHARED_PREF_SOURCE_BYODA_ENABLED, true)) {
-                        if(dexcomReceiver == null) {
-                            val dexcomFilter = IntentFilter()
-                            dexcomFilter.addAction(Intents.DEXCOM_CGM_BROADCAST_ACTION)
-                            dexcomFilter.addAction(Intents.DEXCOM_G7_BROADCAST_ACTION)
-                            dexcomReceiver = DexcomBroadcastReceiver()
-                            if(!registerReceiver(context, dexcomReceiver!!, dexcomFilter))
-                                dexcomReceiver = null
-                        }
-                    } else if (dexcomReceiver != null) {
-                        unregisterReceiver(context, dexcomReceiver)
-                        dexcomReceiver = null
-                    }
-                }
-
-                if(key.isNullOrEmpty() || key == Constants.SHARED_PREF_SOURCE_EVERSENSE_ENABLED) {
-                    if (sharedPref.getBoolean(Constants.SHARED_PREF_SOURCE_EVERSENSE_ENABLED, true)) {
-                        if(nsEmulatorReceiver == null) {
-                            nsEmulatorReceiver = NsEmulatorReceiver()
-                            if(!registerReceiver(context, nsEmulatorReceiver!!, IntentFilter(Intents.NS_EMULATOR_BROADCAST_ACTION)))
-                                nsEmulatorReceiver = null
-                        }
-                    } else if (nsEmulatorReceiver != null) {
-                        unregisterReceiver(context, nsEmulatorReceiver)
-                        nsEmulatorReceiver = null
-                    }
-                }
-
-                if(key.isNullOrEmpty() || key == Constants.SHARED_PREF_SOURCE_DIABOX_ENABLED) {
-                    if (sharedPref.getBoolean(Constants.SHARED_PREF_SOURCE_DIABOX_ENABLED, true)) {
-                        if(diaboxReceiver == null) {
-                            diaboxReceiver = DiaboxReceiver()
-                            if(!registerReceiver(context, diaboxReceiver!!, IntentFilter(Intents.DIABOX_BROADCAST_ACTION)))
-                                diaboxReceiver = null
-                        }
-                    } else if (diaboxReceiver != null) {
-                        unregisterReceiver(context, diaboxReceiver)
-                        diaboxReceiver = null
-                    }
-                }
-
-                if(key.isNullOrEmpty() || key == Constants.SHARED_PREF_SOURCE_LIBRE_PATCHED_ENABLED) {
-                    if (sharedPref.getBoolean(Constants.SHARED_PREF_SOURCE_LIBRE_PATCHED_ENABLED, true)) {
-                        if(librePatchedReceiver == null) {
-                            librePatchedReceiver = LibrePatchedReceiver()
-                            val filter = IntentFilter()
-                            filter.addAction(Constants.XDRIP_ACTION_GLUCOSE_READING)
-                            filter.addAction(Constants.XDRIP_ACTION_SENSOR_ACTIVATE)
-                            if(!registerReceiver(context, librePatchedReceiver!!, filter))
-                                librePatchedReceiver = null
-                        }
-                    } else if (librePatchedReceiver != null) {
-                        unregisterReceiver(context, librePatchedReceiver)
-                        librePatchedReceiver = null
-                    }
-                }
-                if (key.isNullOrEmpty() || key == Constants.SHARED_PREF_SOURCE_AIDEX_ENABLED) {
-                    if (sharedPref.getBoolean(Constants.SHARED_PREF_SOURCE_AIDEX_ENABLED, true)) {
-                        if (aidexReceiver == null) {
-                            aidexReceiver = AidexBroadcastReceiver()
-                            val filter = IntentFilter()
-                            filter.addAction(Intents.AIDEX_BROADCAST_ACTION)
-                            if (!registerReceiver(context, aidexReceiver!!, filter))
-                                aidexReceiver = null
-                        }
-                    } else if (aidexReceiver != null) {
-                        unregisterReceiver(context, aidexReceiver)
-                        aidexReceiver = null
-                    }
-
-                }
-
-                if (key.isNullOrEmpty() || key == Constants.SHARED_PREF_SOURCE_NOTIFICATION_ENABLED) {
-                    updateNotificationReceiver(sharedPref, context)
-                }
-
-
-            } catch (exc: Exception) {
-                Log.e(LOG_ID, "registerSourceReceiver exception: " + exc.toString())
-            }
-        }
-
-        fun checkNotificationReceiverPermission(context: Context, requestPermission: Boolean): Boolean {
-            val notificationListeners = Settings.Secure.getString(context.contentResolver, "enabled_notification_listeners")
-            if(!notificationListeners.contains(context.packageName)) {
-                if(requestPermission) {
-                    // request permissions
-                    val intent = Intent(Settings.ACTION_NOTIFICATION_LISTENER_SETTINGS)
-                    intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-                    context.startActivity(intent)
-                }
-                return false
-            }
-            return true
-        }
-
-        fun updateNotificationReceiver(sharedPref: SharedPreferences, context: Context) {
-            try {
-                // default to false because reading notifications is a scary permission to give for no reason
-                if (sharedPref.getBoolean(Constants.SHARED_PREF_SOURCE_NOTIFICATION_ENABLED, false)) {
-                    Log.i(LOG_ID, "Notification source enabled")
-                    checkNotificationReceiverPermission(context, true)
-                    notificationReceiver = NotificationReceiver()
-                    registerReceiver(context, notificationReceiver!!, IntentFilter())
-                } else if(notificationReceiver!=null) {
-                    Log.i(LOG_ID, "Notification source disabled")
-                    unregisterReceiver(context, notificationReceiver)
-                    notificationReceiver = null
-                }
-                // notification listeners can not be unregistered
-            } catch (exc: Exception) {
-                Log.e(LOG_ID, "updateNotificationReceiver exception: " + exc.toString())
-            }
-        }
-
-        fun unregisterSourceReceiver(context: Context) {
-            try {
-                Log.d(LOG_ID, "Unregister receiver")
-                if (glucoDataReceiver != null) {
-                    unregisterReceiver(context, glucoDataReceiver)
-                    glucoDataReceiver = null
-                }
-                if (xDripReceiver != null) {
-                    unregisterReceiver(context, xDripReceiver)
-                    xDripReceiver = null
-                }
-                if (dexcomReceiver != null) {
-                    unregisterReceiver(context, dexcomReceiver)
-                    dexcomReceiver = null
-                }
-                if (nsEmulatorReceiver != null) {
-                    unregisterReceiver(context, nsEmulatorReceiver)
-                    nsEmulatorReceiver = null
-                }
-                if (aapsReceiver != null) {
-                    unregisterReceiver(context, aapsReceiver)
-                    aapsReceiver = null
-                }
-                if (diaboxReceiver != null) {
-                    unregisterReceiver(context, diaboxReceiver)
-                    diaboxReceiver = null
-                }
-                if (librePatchedReceiver != null) {
-                    unregisterReceiver(context, librePatchedReceiver)
-                    librePatchedReceiver = null
-                }
-                if(notificationReceiver != null) {
-                    unregisterReceiver(context, notificationReceiver)
-                    notificationReceiver = null
-                }
-                if (aidexReceiver != null) {
-                    unregisterReceiver(context, aidexReceiver)
-                    aidexReceiver = null
-                }
-            } catch (exc: Exception) {
-                Log.e(LOG_ID, "unregisterSourceReceiver exception: " + exc.toString())
-            }
-        }
-
-        fun getSettings(): Bundle {
-            val bundle = ReceiveData.getSettingsBundle()
-            // other settings
-            if (sharedPref != null) {
-                bundle.putBoolean(Constants.SHARED_PREF_SHOW_OTHER_UNIT, sharedPref!!.getBoolean(Constants.SHARED_PREF_SHOW_OTHER_UNIT, ReceiveData.isMmol))
-                bundle.putBoolean(Constants.SHARED_PREF_SOURCE_JUGGLUCO_ENABLED, sharedPref!!.getBoolean(Constants.SHARED_PREF_SOURCE_JUGGLUCO_ENABLED, true))
-                bundle.putBoolean(Constants.SHARED_PREF_SOURCE_XDRIP_ENABLED, sharedPref!!.getBoolean(Constants.SHARED_PREF_SOURCE_XDRIP_ENABLED, true))
-                bundle.putBoolean(Constants.SHARED_PREF_SOURCE_AAPS_ENABLED, sharedPref!!.getBoolean(Constants.SHARED_PREF_SOURCE_AAPS_ENABLED, true))
-                bundle.putBoolean(Constants.SHARED_PREF_SOURCE_BYODA_ENABLED, sharedPref!!.getBoolean(Constants.SHARED_PREF_SOURCE_BYODA_ENABLED, true))
-                bundle.putBoolean(Constants.SHARED_PREF_SOURCE_EVERSENSE_ENABLED, sharedPref!!.getBoolean(Constants.SHARED_PREF_SOURCE_EVERSENSE_ENABLED, true))
-                bundle.putBoolean(Constants.SHARED_PREF_SOURCE_DIABOX_ENABLED, sharedPref!!.getBoolean(Constants.SHARED_PREF_SOURCE_DIABOX_ENABLED, true))
-                bundle.putBoolean(Constants.SHARED_PREF_SOURCE_NOTIFICATION_ENABLED, sharedPref!!.getBoolean(Constants.SHARED_PREF_SOURCE_NOTIFICATION_ENABLED, false))
-                bundle.putBoolean(Constants.SHARED_PREF_PHONE_WEAR_SCREEN_OFF_UPDATE, sharedPref!!.getBoolean(Constants.SHARED_PREF_PHONE_WEAR_SCREEN_OFF_UPDATE, true))
-                bundle.putString(Constants.SHARED_PREF_SENSOR_RUNTIME, sharedPref!!.getString(Constants.SHARED_PREF_SENSOR_RUNTIME, "14"))
-                bundle.putString(Constants.PATIENT_NAME, sharedPref!!.getString(Constants.PATIENT_NAME, ""))
-            }
-            if(Log.isLoggable(LOG_ID, android.util.Log.VERBOSE))
-                Log.v(LOG_ID, "getSettings called with bundle ${(Utils.dumpBundle(bundle))}")
-            return bundle
+        fun getSettings(): Bundle? {
+            return service?.getSettings()
         }
 
         fun setSettings(context: Context, bundle: Bundle) {
-            if(Log.isLoggable(LOG_ID, android.util.Log.VERBOSE))
-                Log.v(LOG_ID, "setSettings called with bundle ${(Utils.dumpBundle(bundle))}")
-            val sharedPref = context.getSharedPreferences(Constants.SHARED_PREF_TAG, MODE_PRIVATE)
-            with(sharedPref!!.edit()) {
-                putBoolean(Constants.SHARED_PREF_SHOW_OTHER_UNIT, bundle.getBoolean(Constants.SHARED_PREF_SHOW_OTHER_UNIT, ReceiveData.isMmol))
-                putBoolean(Constants.SHARED_PREF_SOURCE_JUGGLUCO_ENABLED, bundle.getBoolean(Constants.SHARED_PREF_SOURCE_JUGGLUCO_ENABLED, true))
-                putBoolean(Constants.SHARED_PREF_SOURCE_XDRIP_ENABLED, bundle.getBoolean(Constants.SHARED_PREF_SOURCE_XDRIP_ENABLED, true))
-                putBoolean(Constants.SHARED_PREF_SOURCE_AAPS_ENABLED, bundle.getBoolean(Constants.SHARED_PREF_SOURCE_AAPS_ENABLED, true))
-                putBoolean(Constants.SHARED_PREF_SOURCE_BYODA_ENABLED, bundle.getBoolean(Constants.SHARED_PREF_SOURCE_BYODA_ENABLED, true))
-                putBoolean(Constants.SHARED_PREF_SOURCE_EVERSENSE_ENABLED, bundle.getBoolean(Constants.SHARED_PREF_SOURCE_EVERSENSE_ENABLED, true))
-                putBoolean(Constants.SHARED_PREF_SOURCE_DIABOX_ENABLED, bundle.getBoolean(Constants.SHARED_PREF_SOURCE_DIABOX_ENABLED, true))
-                putBoolean(Constants.SHARED_PREF_PHONE_WEAR_SCREEN_OFF_UPDATE, bundle.getBoolean(Constants.SHARED_PREF_PHONE_WEAR_SCREEN_OFF_UPDATE, true))
-                putBoolean(Constants.SHARED_PREF_SOURCE_NOTIFICATION_ENABLED, bundle.getBoolean(Constants.SHARED_PREF_SOURCE_NOTIFICATION_ENABLED, false))
-                putString(Constants.SHARED_PREF_SENSOR_RUNTIME, bundle.getString(Constants.SHARED_PREF_SENSOR_RUNTIME, "14"))
-                putString(Constants.PATIENT_NAME, bundle.getString(Constants.PATIENT_NAME, ""))
-                apply()
-            }
-            ReceiveData.setSettings(sharedPref, bundle)
+            service?.setSettings(context, bundle)
         }
     }
 
@@ -604,13 +204,13 @@ abstract class GlucoDataService(source: AppSource) : WearableListenerService(), 
                 }
                 isForegroundService = true
                 Log.i(LOG_ID, "Service in foreground started!")
-                stopTrigger()
+                StartupTrigger.stopTrigger()
             }
         } catch (exc: Exception) {
             Log.e(LOG_ID, "onStartCommand exception: " + exc.toString())
             if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.S && exc is ForegroundServiceStartNotAllowedException) {
                 if(startServiceReceiver != null) {
-                    triggerStartService(this, startServiceReceiver!!)
+                    StartupTrigger.triggerStartService(this, startServiceReceiver!!)
                 }
             }
         }
@@ -632,10 +232,9 @@ abstract class GlucoDataService(source: AppSource) : WearableListenerService(), 
             SourceTaskService.run(this)
             PackageUtils.updatePackages(this)
 
-            connection = WearPhoneConnection()
-            connection!!.open(this, appSource == AppSource.PHONE_APP)
+            WearPhoneManager.init(this, appSource == AppSource.PHONE_APP)
 
-            updateSourceReceiver(this)
+            ReceiverManager.updateSourceReceiver(this)
             broadcastServiceAPI = BroadcastServiceAPI()
             broadcastServiceAPI.init()
             updateBatteryReceiver()
@@ -674,7 +273,7 @@ abstract class GlucoDataService(source: AppSource) : WearableListenerService(), 
         try {
             Log.w(LOG_ID, "onDestroy called")
             sharedPref!!.unregisterOnSharedPreferenceChangeListener(this)
-            unregisterSourceReceiver(this)
+            ReceiverManager.unregisterSourceReceiver(this)
             broadcastServiceAPI.close(this)
             if(batteryReceiver != null) {
                 unregisterReceiver(batteryReceiver)
@@ -686,8 +285,7 @@ abstract class GlucoDataService(source: AppSource) : WearableListenerService(), 
             }
             TimeTaskService.stop()
             SourceTaskService.stop()
-            connection!!.close()
-            connection = null
+            WearPhoneManager.close()
             super.onDestroy()
             service = null
             created = false
@@ -740,7 +338,7 @@ abstract class GlucoDataService(source: AppSource) : WearableListenerService(), 
                 Constants.SHARED_PREF_SOURCE_DIABOX_ENABLED,
                 Constants.SHARED_PREF_SOURCE_AIDEX_ENABLED,
                 Constants.SHARED_PREF_SOURCE_NOTIFICATION_ENABLED -> {
-                    updateSourceReceiver(this, key)
+                    ReceiverManager.updateSourceReceiver(this, key)
                     shareSettings = true
                 }
                 Constants.SHARED_PREF_SHOW_OTHER_UNIT,
@@ -767,20 +365,12 @@ abstract class GlucoDataService(source: AppSource) : WearableListenerService(), 
         }
     }
 
-    fun sendToConnectedDevices(dataSource: NotifySource, extras: Bundle) {
-        Thread {
-            try {
-                if(connection != null)
-                    connection!!.sendMessage(dataSource, extras, null)
-            } catch (exc: Exception) {
-                Log.e(LOG_ID, "SendMessage exception: " + exc.toString())
-            }
-        }.start()
-    }
-
     override fun onTimeout(startId: Int, fgsType: Int) {
         super.onTimeout(startId, fgsType)
         Log.e(LOG_ID, "onTimeout called with startId $startId and fgsType $fgsType")
         //stopSelf() better force crash to see, what happens...
     }
+
+    abstract fun getSettings(): Bundle?
+    abstract fun setSettings(context: Context, bundle: Bundle)
 }
