+++ common/src/main/java/de/michelinside/glucodatahandler/common/receiver/NamedReceiver.kt
@@ -1,11 +1,11 @@
 package de.michelinside.glucodatahandler.common.receiver
 
-import de.michelinside.glucodatahandler.common.GlucoDataService
+import de.michelinside.glucodatahandler.common.service.ReceiverManager
 
 interface NamedReceiver {
     fun getName(): String
 
     fun isRegistered(): Boolean {
-        return GlucoDataService.isRegistered(this)
+        return ReceiverManager.isRegistered(this)
     }
 }
