+++ mobile/src/main/java/de/michelinside/glucodatahandler/android_auto/CarModeReceiver.kt
@@ -129,8 +129,13 @@ object CarModeReceiver {
                     if(withSettings) {
                         if(sharedPref.getBoolean(Constants.SHARED_PREF_SEND_PREF_TO_GLUCODATAAUTO, true)) {
                             val settings = GlucoDataService.getSettings()
-                            settings.putBoolean(Constants.SHARED_PREF_RELATIVE_TIME, ElapsedTimeTask.relativeTime)
-                            extras.putBundle(Constants.SETTINGS_BUNDLE, settings)
+                            if(settings != null) {
+                                settings.putBoolean(
+                                    Constants.SHARED_PREF_RELATIVE_TIME,
+                                    ElapsedTimeTask.relativeTime
+                                )
+                                extras.putBundle(Constants.SETTINGS_BUNDLE, settings)
+                            }
                             extras.putBundle(Constants.ALARM_SETTINGS_BUNDLE, AlarmHandler.getSettings(true))
                         }
                     }
